Update: Reorganized the file structure. Added changes to Scheduler_FSM and generated a cpp testbench to verify my FSM logic is correct
Description:
Originally I created a cpp testbench that was hanging when waiting for
the compute block to finish. So I changed how we transition between
states within multi-headed attention slightly to account for that. I
also added some requant logic at the top of the file, but it does
nothing right now (for debug purposes).

I also ran the FSM through vitis' HLS compiler and put its Verilog
outputs in the Scheduler_FSM directory under Synthesize-Verilog. I plan
to keep the naming scheme the same for all modules we create.

Things to do next:
Still the FSM is not perfect and requires updates. Here are two features
I need to implement before finalizing:
1. Make Requenat an external Compute Operation, seperate from the MAC
   compute
2. Implement URAM/BRAM load requests instead of straight from DMA.
   Context: At the moment, for each compute step we load the head/tile
but then we don't tell the compute module where to actually pull the
data from. I plan on making some kind of Memory Data Manager that stores
aliases for all data stored in the URAM, so the compute block can
request for a stage, state or type of data and properly be fed input.