Subject: Updating broken testbench
Description:
The challenge right now is making a testbench that can handle data in
parallel. Currently the issue is that if we have two parallel heads in a
group, then we have one head dominate the resources provided by the
System Verilog testbench. This causes one of the heads to stay idle at
the first head state that requires compute requests (Q). Trying to get
blocks to trigger at the same time. Saving current progress, so as to
not degrade from current version.